pyproject.toml
[tool.poetry]
name = "whisper-writer-2"
version = "0.1.0"
description = ""
authors = ["N <47500890+avi-cenna@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
pyzmq = "^25.1.1"
torch = "^2.1.0"
openai-whisper = {git = "https://github.com/openai/whisper.git"}
openai = "^0.28.1"
keyboard = "^0.13.5"
pyautogui = "^0.9.54"
sounddevice = "^0.4.6"
webrtcvad = "^2.0.10"
python-dotenv = "^1.0.0"
black = "^23.10.1"
isort = "^5.12.0"
pynput = "^1.7.6"
loguru = "^0.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
